backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pdfplumber==0.9.0
python-docx==0.8.11
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
Werkzeug==2.3.7